规格绑定分类,分类绑定品牌和规格,不能被删除

+++ xmd-shop-parent/xmd-shop-service/xmd-shop-service-xxx/src/main/java/com/baidu/shop/service/impl/CategoryServiceImpl.java
@@ -2,8 +2,14 @@ package com.baidu.shop.service.impl;
 
 import com.baidu.shop.base.BaseApiService;
 import com.baidu.shop.base.Result;
+import com.baidu.shop.entity.CategoryBrandEntity;
 import com.baidu.shop.entity.CategoryEntity;
+import com.baidu.shop.entity.SpecGroupEntity;
+import com.baidu.shop.entity.SpecParamEntity;
+import com.baidu.shop.mapper.CategoryBrandMapper;
 import com.baidu.shop.mapper.CategoryMapper;
+import com.baidu.shop.mapper.SpecGroupMapper;
+import com.baidu.shop.mapper.SpecParamMapper;
 import com.baidu.shop.service.CategoryService;
 import com.google.gson.JsonObject;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,6 +31,12 @@ public class CategoryServiceImpl extends BaseApiService implements CategoryServi
 
     @Resource
     private CategoryMapper categoryMapper;
+
+    @Resource
+    private SpecGroupMapper specGroupMapper;
+
+    @Resource
+    private CategoryBrandMapper categoryBrandMapper;
 
 
     @Override
@@ -80,6 +92,7 @@ public class CategoryServiceImpl extends BaseApiService implements CategoryServi
             return this.setResultSuccess("当前节点为父节点 不能被删除");
         }
 
+
         Example example = new Example(CategoryEntity.class);
         example.createCriteria().andEqualTo("parentId",categoryEntity.getParentId());
         List<CategoryEntity> list =  categoryMapper.selectByExample(example);
@@ -89,6 +102,24 @@ public class CategoryServiceImpl extends BaseApiService implements CategoryServi
             parentCateEntity.setId(categoryEntity.getParentId());
             parentCateEntity.setIsParent(0);
             categoryMapper.updateByPrimaryKeySelective(parentCateEntity);
+        }
+
+
+        //分类绑定规格组 不能删除
+        Example example1 = new Example(SpecGroupEntity.class);
+        example1.createCriteria().andEqualTo("cid",id);
+        List<SpecGroupEntity> list1 = specGroupMapper.selectByExample(example1);
+        if(list1.size() == 1){
+            return this.setResultError("分类绑定规格组不能删除");
+        }
+
+
+        //分类绑定品牌
+        Example example2 = new Example(CategoryBrandEntity.class);
+        example2.createCriteria().andEqualTo("categoryId",id);
+        List<CategoryBrandEntity> list2 = categoryBrandMapper.selectByExample(example2);
+        if(list2.size() == 1){
+            return this.setResultError("分类绑定品牌不能删除");
         }
 
         categoryMapper.deleteByPrimaryKey(id);
